ISSUE #1544 - Now you can attach both files and urls at the same time
